excludes: Undo shell-style patterns, use relative paths to $HOME instead (after a99ccd9)
borgmatic interprets source directories relative to the CWD, so we need to chdir to $HOME before creating a backup to be able to use relative paths.
Relative paths are important to be able to reuse the excludes for the userdata-migrater; the previous change broke that script!

(This corrects commit a99ccd976693ef4defc550e4e690a4f6d2ead574)